initial scaffolding of routes and layout
