+++ package.json
@@ -15,7 +15,7 @@
     "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router": "^7.0.1"
+    "react-router": "^7.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
