+++ src/pages/PlaylistPage.tsx
@@ -1,0 +1,3 @@
+export default function PlaylistPage() {
+  return <h1>Playlist</h1>;
+}
